+++ RELEASENOTES.md
@@ -26,6 +26,8 @@
     *   Remove `FileDataSourceFactory`. Use `FileDataSource.Factory` instead.
     *   Remove `SimpleExoPlayer.addMetadataOutput` and `removeMetadataOutput`.
         Use `Player.addListener` and `Player.Listener` instead.
+    *   Remove `SimpleExoPlayer.addAudioListener`, `removeAudioListener` and
+        `AudioListener`. Use `Player.addListener` and `Player.Listener` instead.
 
 ### 2.15.0 (2021-08-10)